# openapi-herramientas-equipos.yaml
openapi: 3.0.3
info:
  title: API de Herramientas y Equipos Retirados
  description: >
    Módulo de backend de la app Wifix (Fase 1). Cubre dos dominios del
    servicio de operaciones de campo:

      1. HERRAMIENTAS — almacenamiento de resultados de mediciones:
         medición de distancia, test de velocidad, mapa de calor WiFi,
         ping y traceroute.
      2. EQUIPOS RETIRADOS — registro de equipos retirados o reemplazados.

    CLAVE DE IDENTIFICACIÓN (FASE 1)
    ----------------------------------------------------------------------
    En la Fase 1 la clave práctica de cada registro es `accountNumber`:
    el número de cuenta del cliente que el técnico ingresa en la app.
    Los demás identificadores (`clientId`, `visitId`, `technicianId`) son
    opcionales: provendrán de un sistema upstream aún por definir y en la
    Fase 1 se completan con valores de prueba. `contractId` también es
    opcional y en una fase posterior será reemplazado por `taskId`.

    AUTENTICACIÓN
    ----------------------------------------------------------------------
    La app tiene login propio. POST /auth/login devuelve un JWT que el
    resto de endpoints exige en el header Authorization: Bearer.

    DATOS DE SISTEMAS EXTERNOS (MOCKUP)
    ----------------------------------------------------------------------
    Los campos que provienen de sistemas de la operadora (Comarch/TYTAN,
    FSM, ISP Monitor, ACS, TEC, RMS) se sirven en esta etapa con datos
...
  version: 1.0.0
  contact:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /naps/nearby:
    get:
      tags: [Diagnóstico de Red]
      summary: NAPs cercanas a una coordenada
      description: >
        NAPs cercanas con coordenada, distancia en metros y puertos
        ocupados/libres (campo 6). La consulta es por coordenada (GPS del
        técnico o de la tarea), no por número de cuenta, igual que la
        pestaña "GPON" de FSM. **1 llamada upstream.**


        La fuente la decide `NAPS_PRIMARY_SOURCE` (ADR-04): FSM
        `GET /naps/nearest` o TEC `GET /api/tec/naps/{lat},{lng}` (Digest).
        Con FSM primario, si el token de la operadora falla se cae a TEC y se
        responde 200 con `source:"TEC"` y el aviso en el header
        `X-Wifix-Degraded`. Cualquier otro fallo de FSM se propaga como 502.


        ⚠ La respuesta es un **array desnudo**: por eso el bloque `degraded`
        viaja en el header `X-Wifix-Degraded` (JSON URL-encoded) y no en el
        cuerpo.
      operationId: getNearbyNaps
      parameters:
        - name: lat
          in: query
          required: true
          description: Latitud desde la que se busca.
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
        - name: lng
          in: query
          required: true
          description: Longitud desde la que se busca.
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
        - name: meters
          in: query
          required: false
          description: Radio de búsqueda en metros. Se ignora si la fuente efectiva es TEC.
          schema:
            type: integer
            minimum: 1
            maximum: 2000
            default: 100
        - name: maxRows
          in: query
          required: false
          description: Cantidad máxima de NAPs. Se ignora si la fuente efectiva es TEC.
          schema:
            type: integer
            minimum: 1
            maximum: 25
            default: 5
        - $ref: '#/components/parameters/BrandHeader'
        - $ref: '#/components/parameters/BrandQuery'
      responses:
        '200':
          description: OK. Lista ordenada por distancia ascendente (vacía si no hay NAPs en el sector).
          headers:
            X-Wifix-Degraded:
              description: >
                Presente solo si la respuesta se sirvió con una fuente
                alternativa o incompleta. Contiene el JSON del bloque
                `Degraded` URL-encoded. El cliente nunca debe asumir que existe.
              schema:
                type: string
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NearbyNap'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          $ref: '#/components/responses/ConnectorError'
        '503':
          $ref: '#/components/responses/UpstreamAuthUnavailable'
  /naps/{napRef}/ports:
    get:
      tags: [Diagnóstico de Red]
      summary: Puertos ocupados por NAP (campo 8, paso 1)
      description: >
        Rejilla de puertos de la NAP con la cuenta y el equipo GPON de cada
        puerto ocupado (campo 8). **1 sola llamada upstream**
        (`GET /naps/accounts`): los estados A/S/T/O/P NO se consultan acá.
        Cada puerto ocupado viene con `clientStatus: null` y
        `statusPending: true`; el técnico los resuelve después, de forma
        explícita, con `POST /accounts/status-batch` (ADR-05).


        **Ruteo del path param**: si `napRef` es numérico (`/^\d+$/`) se toma
        como `napId` de FSM; cualquier otro valor se toma como código de NAP y
        va por el camino TEC heredado, que sigue devolviendo
        `detailAvailable:false`. El frontend envía `nap.napId ?? nap.napCode`.
      operationId: getNapPorts
      parameters:
        - name: napRef
          in: path
          required: true
          description: >
            Id numérico de la NAP en FSM, o código de NAP (camino TEC).
          schema:
            type: string
            example: '11547'
        - name: withStatus
          in: query
          required: false
          description: >
            Resuelve además hasta `batchLimit` estados en la misma llamada.
            Existe SOLO para `probe-fsm-api.ts` y soporte: **la webapp tiene
            prohibido enviarlo**, porque dispara hasta 12 consultas extra a
            producción por cada tap.
          schema:
            type: string
            enum: ['0', '1']
            default: '0'
        - $ref: '#/components/parameters/BrandHeader'
        - $ref: '#/components/parameters/BrandQuery'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NapPorts'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/ConnectorError'
        '503':
          $ref: '#/components/responses/UpstreamAuthUnavailable'
  /terminals/{id}:
    get:
      tags: [Diagnóstico de Red]
      summary: Estado del equipo, de la red y evento asociado
      description: >
        Ficha puntual del equipo del cliente: estado del terminal, estado de
        la red y evento (daño) asociado. `id` es el serial GPON (fibra) o la
        MAC del cablemódem (HFC); las MAC se aceptan con o sin separadores.
        Origen real: ISP Monitor (`GET /api/isp/terminals/{id}`).
      operationId: getTerminal
      parameters:
        - $ref: '#/components/parameters/TerminalIdPath'
      responses:
        '200':
          description: >
            OK. Si el identificador no existe en ISP Monitor, `found` es false
            (la API de operadora responde 204 en ese caso).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalSnapshot'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /terminals/{id}/diagnostics:
    get:
      tags: [Diagnóstico de Red]
      summary: Panel completo de ISP Monitor (campos 9-12)
      description: >
        Ficha del equipo más las series de las últimas 24 horas (estado, señal
        a ruido y errores FEC, cada una para el terminal y para la red) en una
        sola llamada.

        Solo se consulta lo que aplica: primero la ficha y, según la tecnología
        que devuelva, las series. En GPON las cuatro series DOCSIS no se piden
        (la operadora responde 204) y en un identificador inexistente no se pide
        ninguna, así que el panel cuesta 3 llamadas upstream en fibra y 7 en
        HFC. Lo omitido se reporta en `skipped`.

        Las series que sí aplican se consultan en paralelo, con un tope global
        de concurrencia y un cache corto (`TEC_API_MAX_CONCURRENCY`,
        `TEC_API_CACHE_TTL_MS`): la operadora no impone un límite de peticiones
        pero pidió no abusar, y todas las consultas van contra producción. Si
        alguna falla, el resto igual se devuelve y el fallo se reporta en
        `errors`.
      operationId: getTerminalDiagnostics
      parameters:
        - $ref: '#/components/parameters/TerminalIdPath'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TerminalDiagnostics'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /terminals/{id}/series/{scope}/{metric}:
    get:
      tags: [Diagnóstico de Red]
      summary: Serie de 24 horas de una métrica
      description: >
        Serie individual de las últimas 24 horas. Combinaciones y endpoint
        upstream correspondiente:
        `terminal/status` → `/api/isp/status/{id}`;
        `network/status` → `/api/isp/network/online/{id}`;
        `terminal/snr` → `/api/isp/cablemodem/snr/{id}`;
        `network/snr` → `/api/isp/network/snr/{id}`;
        `terminal/codewords` → `/api/isp/cablemodem/codewords/{id}`;
        `network/codewords` → `/api/isp/network/codewords/{id}`.
      operationId: getTerminalSeries
      parameters:
        - $ref: '#/components/parameters/TerminalIdPath'
        - name: scope
          in: path
          required: true
          description: Ámbito de la serie — el equipo del cliente o la red/nodo.
          schema:
            type: string
            enum: [terminal, network]
        - name: metric
          in: path
          required: true
          description: >
            Métrica: `status` (disponibilidad, campo 12), `snr` (señal a
            ruido, campo 10) o `codewords` (FEC corregidos y sin corregir,
            campo 11).
          schema:
            type: string
            enum: [status, snr, codewords]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Series24h'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /accounts/{accountNumber}/node-events:
    get:
      tags: [Diagnóstico de Red]
      summary: Daños (eventos) en la red de acceso
      description: >
        Eventos o daños de la red que da servicio a la cuenta (campo 14).
        Origen real: RMS — todavía en mock.

        La ruta, el `operationId` y el esquema conservan el término "node" del
        Excel original y del contrato ya publicado. En pantalla se dice "red de
        acceso": la operadora aclaró (2026-08-27) que "nodo" no es un concepto
        de su red — sus datos salen de tarjetas de CMTS o de puertos de OLT.
        Cuando haya credenciales de RMS habrá que confirmar qué agrupa
        realmente este endpoint.
      operationId: getNodeEvents
      parameters:
        - $ref: '#/components/parameters/AccountNumberPath'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NodeEvent'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /accounts/{accountNumber}/lan-devices:
    get:
      tags: [Diagnóstico de Red]
      summary: Equipos en la red local (tabla DHCP)
      description: >
        Lista de equipos conectados en la red local del cliente
        (campo 19). Origen real: ISP / ACS.
      operationId: getLanDevices
      parameters:
        - $ref: '#/components/parameters/AccountNumberPath'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LanDevice'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /accounts/{accountNumber}/wifi-devices:
    get:
      tags: [Diagnóstico de Red]
      summary: Dispositivos conectados por WiFi
      description: >
        Dispositivos conectados por WiFi en cada banda (2.4 y 5 GHz) con
        su nivel de señal (campo 20). Origen real: ISP / ACS.
      operationId: getWifiDevices
      parameters:
        - $ref: '#/components/parameters/AccountNumberPath'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WifiDevice'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /accounts/{accountNumber}/wifi-config:
    parameters:
      - $ref: '#/components/parameters/AccountNumberPath'
    get:
      tags: [Diagnóstico de Red]
      summary: Configuración WiFi actual
      operationId: getWifiConfig
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WifiConfig'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Diagnóstico de Red]
      summary: Cambiar SSID y contraseña
      description: >
        Cambia el SSID y la contraseña del WiFi del cliente (campo 21).
        Operación sensible. Origen real: ACS.
      operationId: updateWifiConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WifiConfigUpdate'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WifiConfig'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ====================================================================
  # TAREAS Y VISITAS (campos 15-16)
  # ====================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT emitido por POST /auth/login.

  # --------------------------------------------------------------------
  # PARÁMETROS REUTILIZABLES
  # --------------------------------------------------------------------
  parameters:
...
    AccountNumberPath:
      name: accountNumber
      in: path
      required: true
      description: Número de cuenta del cliente.
      schema:
        type: string
    TerminalIdPath:
      name: id
      in: path
      required: true
      description: >
        Serial GPON (fibra) o MAC del cablemódem (HFC). Las MAC se aceptan
        con o sin separadores; el backend las normaliza a hexadecimal plano
        en mayúsculas porque IIS rechaza los ":" en la ruta.
      schema:
        type: string
        example: ZTEGC1234567
    BrandHeader:
      name: X-Wifix-Brand
      in: header
      required: false
      description: >
        Marca comercial (realm de FSM) con la que se consulta la operadora.
        Si falta, el backend usa `FSM_DEFAULT_BRAND`. Internamente nunca se
        habla de "realm": los nombres `realm-ecommerce-callcenter-*` quedan
        encerrados en el backend y el técnico no los ve (ADR-01).
      schema:
        type: string
        enum: [telenews, seteinfo]
        example: telenews
    BrandQuery:
      name: brand
      in: query
      required: false
      description: >
        Override de `X-Wifix-Brand` por query. Pensado para el script de sondeo
        y soporte; la webapp usa el header.
      schema:
        type: string
        enum: [telenews, seteinfo]
...
  responses:
    BadRequest:
      description: Solicitud inválida.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Token ausente, inválido o expirado / credenciales incorrectas.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Recurso no encontrado.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ConnectorError:
      description: >
        Un sistema externo falló (5xx, timeout o JSON inválido). No es un
        problema de credenciales.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    UpstreamAuthUnavailable:
      description: >
        El acceso a FSM no está disponible: token ausente, vencido o rechazado.


        **Es 503 a propósito, nunca 401.** La webapp cierra la sesión del
        técnico ante cualquier 401 (`api.js`), y un token caducado de un tercero
        no puede sacarlo al login en medio de una visita ni tumbarle Herramientas
        y Equipos Retirados, que no dependen de FSM.


        El cliente debe pintar un banner no bloqueante en el panel afectado y
        **no** cerrar sesión, no limpiar el token de Wifix y no recargar la app.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UpstreamAuthError'

  # --------------------------------------------------------------------
  # ESQUEMAS
  # --------------------------------------------------------------------
  schemas:
...
    NearbyNap:
      type: object
      description: NAP cercana (campo 6). La fuente la indica `source`.
      properties:
        napId:
          type: integer
          nullable: true
          description: >
            Id numérico de la NAP en FSM. Es la entrada de
            `GET /naps/{napRef}/ports`. `null` cuando la fuente es TEC.
          example: 11547
        napCode:
          type: string
          example: PL2KD9
        networkName:
          type: string
          nullable: true
          description: >
            Red de acceso (OLT/puerto) de la que cuelga la NAP. Nunca se rotula
            como "nodo": la operadora confirmó que ese concepto no existe.
            `null` cuando la fuente es TEC.
          example: OLT-GYE-03/1/2
        latitude:
          type: number
          format: double
          nullable: true
          description: Coordenada de la NAP; null si la operadora no la devuelve.
        longitude:
          type: number
          format: double
          nullable: true
        distanceMeters:
          type: number
          format: double
          description: Distancia calculada por la operadora desde la coordenada consultada.
        occupiedPorts:
          type: integer
        totalPorts:
          type: integer
        freePorts:
          type: integer
          description: totalPorts - occupiedPorts, nunca negativo.
        source:
          type: string
          enum: [FSM, TEC]
          description: De dónde salió la fila (ver `NAPS_PRIMARY_SOURCE`, ADR-04).
...
    NapPorts:
      type: object
      properties:
        napRef:
          type: string
          description: Valor tal como llegó en la ruta (id numérico o código de NAP).
        napId:
          type: integer
          nullable: true
        napCode:
          type: string
          nullable: true
        ports:
          type: array
          description: Vacío cuando `detailAvailable` es false (camino TEC).
          items:
            $ref: '#/components/schemas/NapPort'
        detailAvailable:
          type: boolean
          description: >
            false por el camino TEC: esa API solo publica el conteo
            ocupados/total en el listado de NAPs (el detalle vive en
            tec.grupotvcable.com/Gpon/Coverage, que aún no es API). Con FSM y un
            `napId` numérico es true.
        occupiedPorts:
          type: integer
        totalPorts:
          type: integer
          description: >
            Se toma del `ports` de `/naps/nearest` memorizado. Si no se puede
            determinar, se devuelven solo los puertos conocidos y viene
            `degraded.reason: TRUNCATED`.
        note:
          type: string
          description: Explicación para mostrar al técnico cuando no hay detalle.
        statusFanOut:
          type: object
          description: Cuántos estados quedan por resolver y con qué tope (paso 2).
          properties:
            supported:
              type: boolean
              description: false por el camino TEC.
            pendingAccounts:
              type: integer
            batchLimit:
              type: integer
              example: 12
        source:
          type: string
          enum: [FSM, TEC]
        degraded:
          $ref: '#/components/schemas/Degraded'
...
    TerminalSnapshot:
      type: object
      description: >
        Estado puntual del equipo y evento asociado. Shape verificado contra la
        API real el 2026-08-26 con un ONT ZTE activo.
      properties:
        id:
          type: string
          example: ZTEGD3F9BBE5
        found:
          type: boolean
          description: false cuando ISP Monitor no tiene datos para ese identificador (204).
        online:
          type: boolean
          nullable: true
          description: Estado del equipo del cliente; sale del campo `status` de la API (`up` / `down`).
        technology:
          type: string
          nullable: true
          enum: [GPON, HFC]
        city:
          type: string
          nullable: true
          example: Quito
        networkIds:
          type: array
          description: >
            Identificadores de la red de acceso a la que cuelga el equipo. NO es
            un nodo: la operadora aclaró (2026-08-27) que estos datos se toman
            de una tarjeta de CMTS (HFC — puede cubrir un ramal, un nodo o una
            combinación) o de un puerto de OLT (GPON — un hilo de fibra).
          items:
            type: integer
          example: [9198]
        event:
          type: object
          nullable: true
          description: >
            Evento (daño) asociado. `active` es false cuando el campo `events`
            de la API viene vacío o null.
          properties:
            active:
              type: boolean
            description:
              type: string
              nullable: true
        drop:
          type: object
          description: >
            Caída de red detectada por el sistema de monitoreo de la operadora
            (campo `drop` de la API). Es el único campo "extra" de la ficha que
            la operadora confirmó relevante (2026-08-27); `device`, `ifIndex` e
            `index` son internos del monitoreo y no se exponen en `fields`.
          properties:
            detected:
              type: boolean
            description:
              type: string
              nullable: true
        history:
          type: array
          items:
            $ref: '#/components/schemas/TerminalHistoryEntry'
        fields:
          type: array
          description: >
            Resto de la ficha aplanada, sin repetir lo que ya tiene campo propio
            ni los campos internos del monitoreo (`device`, `ifIndex`, `index`),
            que siguen disponibles en `raw`.
          items:
            $ref: '#/components/schemas/FlatField'
        raw:
          description: Payload original de la API de operadora.
          nullable: true
        fetchedAt:
          type: string
          format: date-time
...
    Series24h:
      type: object
      description: >
        Serie de las últimas 24 horas ya normalizada. La operadora la devuelve
        como tuplas `[[epochSegundos, valor], …]` — 288 muestras, una cada 5
        minutos — sin nombrar las columnas; el backend les pone nombre según la
        métrica y el ámbito:
        `terminal/status` → `online` (1 = en línea, 0 = caído);
        `network/status` → `terminalsOnline` (cuántos equipos de la misma red de
        acceso están en línea, **no** un 0/1; la operadora no publica el total de
        la red, así que no se puede derivar un porcentaje);
        `snr` → `snr` o `snrDown`/`snrUp` según cuántas columnas traiga;
        `codewords` → `errors` o `corrected`/`uncorrected`.
        Las métricas DOCSIS llegan además desglosadas por canal upstream: la
        respuesta es un array de canales, cada uno con su `data` — ver
        `channels`.
        La ventana es siempre de 24 horas contadas desde el instante de la
        consulta; la cantidad de muestras varía según el equipo (se observaron
        entre 122 y 292), así que los promedios deben ponderarse por tiempo, no
        por muestra.
        Las series de SNR y codewords son DOCSIS: en equipos GPON la operadora
        responde 204, por lo que `/diagnostics` directamente no las consulta
        (ver `skipped`).
        Si el formato no se reconoce, `recognized` es false y el consumidor
        puede caer a `raw`.
      properties:
        id:
          type: string
        scope:
          type: string
          enum: [terminal, network]
        metric:
          type: string
          enum: [status, snr, codewords]
        keys:
          type: array
          description: Nombres de las series numéricas presentes (leyenda del gráfico).
          items:
            type: string
        points:
          type: array
          items:
            $ref: '#/components/schemas/SeriesPoint'
        channels:
          type: array
          description: >
            Canales cuando la métrica viene desglosada (SNR y codewords DOCSIS
            traen una serie por canal upstream). Vacío si la métrica es de un
            solo canal; en ese caso `keys`/`points` son la serie. Cuando hay
            canales, `keys` y `points` reflejan el primero, para consumidores
            que no los manejen.
          items:
            $ref: '#/components/schemas/SeriesChannel'
        recognized:
          type: boolean
        raw:
          description: Payload original de la API de operadora.
          nullable: true
        fetchedAt:
          type: string
          format: date-time
...
    TerminalDiagnostics:
      type: object
      description: >
        Ficha del equipo más las series de 24 horas que apliquen (campos 9-12).
        No siempre son seis: ver `skipped`.
      properties:
        id:
          type: string
        terminal:
          $ref: '#/components/schemas/TerminalSnapshot'
        status:
          $ref: '#/components/schemas/ScopedSeries'
        snr:
          $ref: '#/components/schemas/ScopedSeries'
        codewords:
          $ref: '#/components/schemas/ScopedSeries'
        errors:
          type: array
          description: Endpoints upstream que fallaron; el resto del payload sigue siendo válido.
          items:
            type: object
            properties:
              endpoint:
                type: string
                example: network/snr
              message:
                type: string
        skipped:
          type: array
          description: >
            Endpoints que deliberadamente NO se consultaron, y por qué. La
            operadora pidió no consultar datos "sin definir cuándo hacen falta"
            y no existe ambiente de pruebas: en GPON las cuatro series DOCSIS
            (SNR y codewords, terminal y red) no se piden, y si el identificador
            no existe no se pide ninguna serie.
          items:
            type: object
            properties:
              endpoint:
                type: string
                example: terminal/snr
              reason:
                type: string
        window:
          type: object
          description: Ventana que cubren las series, siempre 24 h hacia atrás desde `until`.
          properties:
            hours:
              type: integer
              enum: [24]
            until:
              type: string
              format: date-time
        fetchedAt:
          type: string
          format: date-time
...
    NodeEvent:
      type: object
      description: >
        Evento o daño de la red de acceso que da servicio a la cuenta
        (campo 14). El nombre del esquema conserva el término del contrato
        publicado; ver la nota en `/accounts/{accountNumber}/node-events`.
      properties:
        type:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [RESUELTO, PENDIENTE, FALLA]
        occurredAt:
          type: string
          format: date-time
    LanDevice:
      type: object
      description: Equipo en la tabla DHCP (campo 19).
      properties:
        hostname:
          type: string
        ipAddress:
          type: string
        macAddress:
          type: string
        leaseExpiresAt:
          type: string
          format: date-time
...